syntax: make Pos.offs 32 bits
Offset is arguably the most important of the three.

name                   old time/op    new time/op    delta
Parse/LongStrs-4         5.12µs ± 1%    5.11µs ± 1%    ~     (p=0.699 n=6+6)
Parse/Cmds+Nested-4      7.04µs ± 1%    7.01µs ± 1%    ~     (p=0.177 n=5+6)
Parse/Vars+Clauses-4     5.67µs ± 1%    5.69µs ± 1%    ~     (p=0.485 n=6+6)
Parse/Binary+Redirs-4    6.70µs ± 1%    6.64µs ± 1%    ~     (p=0.132 n=6+6)

name                   old alloc/op   new alloc/op   delta
Parse/LongStrs-4         1.32kB ± 0%    1.38kB ± 0%  +5.17%  (p=0.002 n=6+6)
Parse/Cmds+Nested-4      5.76kB ± 0%    5.98kB ± 0%  +3.75%  (p=0.002 n=6+6)
Parse/Vars+Clauses-4     3.30kB ± 0%    3.47kB ± 0%  +5.21%  (p=0.002 n=6+6)
Parse/Binary+Redirs-4    3.95kB ± 0%    4.14kB ± 0%  +4.86%  (p=0.002 n=6+6)

+++ syntax/parser.go
@@ -111,7 +111,7 @@ func (p *Parser) reset() {
 }
 
 func (p *Parser) getPos() Pos {
-	p.npos.offs = uint16(p.offs + p.bsp - 1)
+	p.npos.offs = uint32(p.offs + p.bsp - 1)
 	return p.npos
 }
